% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bs.R
\name{combine_bs}
\alias{combine_bs}
\title{Combine bootstrap covariance estimate results into estimates of the full covariance matrices}
\usage{
combine_bs(bs_result, limmbo_out)
}
\arguments{
\item{bs_result}{output of function bs_covar_est(); contains the bootstrap covariance estimates}

\item{limmbo_out}{object produced by call to limmbo$core$vdbootstrap$LiMMBo}
}
\value{
a list
}
\description{
Combine bootstrap covariance estimate results into estimates of the full covariance matrices
}
