% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bs.R
\name{convert_for_bscombine}
\alias{convert_for_bscombine}
\title{Helper function to convert for use with combine_bs()}
\usage{
convert_for_bscombine(x)
}
\arguments{
\item{x}{outputted object from bs_covar_est}
}
\description{
Helper function to convert for use with combine_bs()
}
