% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmbo2.R
\name{limmbo2}
\alias{limmbo2}
\title{Estimate Vg and Ve with limmbo}
\usage{
limmbo2(kinship, pheno, niter = 10, S = 2, seed)
}
\arguments{
\item{kinship}{a n by n kinship matrix}

\item{pheno}{a n by d phenotypes matrix}

\item{niter}{integer number of iterations for limix calculations}

\item{S}{number of phenotypes per bootstrap sample}

\item{seed}{positive integer random seed}
}
\description{
Estimate Vg and Ve with limmbo
}
