% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{make_limmbo}
\alias{make_limmbo}
\title{Make a limmbo output object}
\usage{
make_limmbo(datainput, timing, iterations, subset_size)
}
\arguments{
\item{datainput}{outputted object from prep_data}

\item{timing}{logical}

\item{iterations}{positive integer}

\item{subset_size}{positive integer, number of phenotypes per bootstrap sample}
}
\description{
Make a limmbo output object
}
\examples{
\dontrun{
prep_data(pheno = matrix(rnorm(300), nrow = 100), kinship = diag(100)) -> dat
make_limmbo(dat, timing = TRUE, iterations = 10, subset_size = 2)
}
}
