% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for use with bootstrap covariance estimation function}
\usage{
prep_data(pheno, kinship)
}
\arguments{
\item{pheno}{phenotypes matrix}

\item{kinship}{kinship matrix}
}
\description{
Prepare data for use with bootstrap covariance estimation function
}
