% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_covar_est.R
\name{bs_covar_est}
\alias{bs_covar_est}
\title{Perform covariance estimates for each of multiple bootstrap samples of traits}
\usage{
bs_covar_est(limmbo_out, cpus, seed)
}
\arguments{
\item{limmbo_out}{object produced by call to limmbo$core$vdbootstrap$LiMMBo}

\item{cpus}{number of cpus to use (as an R object)}

\item{seed}{integer for pseudo random number seed (as an R object)}
}
\description{
Perform covariance estimates for each of multiple bootstrap samples of traits
}
\examples{
prep_data(pheno = matrix(rnorm(300), nrow = 100), kinship = diag(100)) -> dat
make_limmbo(dat, timing = TRUE, iterations = 10, subset_size = 2) -> limmbo_out
bs_covar_est(limmbo_out, cpus = 1, seed = 100)
}
