% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bs.R
\name{combine_bs}
\alias{combine_bs}
\title{Combine bootstrap covariance estimate results into estimates of the full covariance matrices}
\usage{
combine_bs(bs_result, limmbo_out)
}
\arguments{
\item{bs_result}{output of function bs_covar_est(); contains the bootstrap covariance estimates}

\item{limmbo_out}{object produced by call to limmbo$core$vdbootstrap$LiMMBo}
}
\value{
a list
}
\description{
Combine bootstrap covariance estimate results into estimates of the full covariance matrices
}
\examples{
prep_data(pheno = matrix(rnorm(300), nrow = 100), kinship = diag(100)) -> dat
make_limmbo(dat, timing = TRUE, iterations = 10, subset_size = 2) -> limmbo_out
bs_covar_est(limmbo_out, cpus = 1, seed = 100) -> bs_out
bs_out2 <- lapply(X = bs_out, FUN = convert_for_bscombine)
combine_bs(bs_out2, limmbo_out)
}
