% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{make_limmbo}
\alias{make_limmbo}
\title{Make a limmbo output object}
\usage{
make_limmbo(datainput, timing, iterations, subset_size)
}
\arguments{
\item{datainput}{outputted object from InputData()}

\item{timing}{logical}

\item{iterations}{positive integer}

\item{subset_size}{positive integer, number of phenotypes per bootstrap sample}
}
\description{
Make a limmbo output object
}
